# Make estimateFee faster for large wallets