--- a/lib/core/src/Cardano/Wallet.hs
+++ b/lib/core/src/Cardano/Wallet.hs
@@ -102,7 +102,7 @@ module Cardano.Wallet
     -- ** Payment
     , getTxExpiry
     , selectAssets
-    , readUTxOIndex
+    , readWalletUTxOIndex
     , selectAssetsNoOutputs
     , assignChangeAddresses
     , selectionToUnsignedTx
@@ -1244,17 +1244,17 @@ selectionToUnsignedTx sel s =
             amount = view #coin   bundle
             assets = view #tokens bundle
 
--- | Helper funcion for selectAssets.
-readUTxOIndex
+-- | Read a wallet checkpoint and index its UTxO, for 'selectAssets' and
+-- 'selectAssetsNoOutputs'.
+readWalletUTxOIndex
     :: forall ctx s k. HasDBLayer s k ctx
     => ctx
     -> WalletId
-    -> ExceptT ErrNoSuchWallet IO (WalletId, UTxOIndex, Wallet s, Set Tx)
-readUTxOIndex ctx wid = do
+    -> ExceptT ErrNoSuchWallet IO (UTxOIndex, Wallet s, Set Tx)
+readWalletUTxOIndex ctx wid = do
     (cp, _, pending) <- readWallet @ctx @s @k ctx wid
-    let utxo :: UTxOIndex
-        utxo = UTxOIndex.fromUTxO $ availableUTxO @s pending cp
-    return (wid, utxo, cp, pending)
+    let utxo = UTxOIndex.fromUTxO $ availableUTxO @s pending cp
+    return (utxo, cp, pending)
 
 selectAssetsNoOutputs
     :: forall ctx s k result.
@@ -1264,11 +1264,12 @@ selectAssetsNoOutputs
         , HasNetworkLayer ctx
         )
     => ctx
-    -> (WalletId, UTxOIndex, Wallet s, Set Tx)
+    -> WalletId
+    -> (UTxOIndex, Wallet s, Set Tx)
     -> TransactionCtx
     -> (s -> SelectionResult TokenBundle -> result)
     -> ExceptT ErrSelectAssets IO result
-selectAssetsNoOutputs ctx w@(wid, _, _, _) tx transform = do
+selectAssetsNoOutputs ctx wid wal tx transform = do
     -- NOTE:
     -- Could be made nicer by allowing 'performSelection' to run with no target
     -- outputs, but to satisfy a minimum Ada target.
@@ -1282,7 +1283,7 @@ selectAssetsNoOutputs ctx w@(wid, _, _, _) tx transform = do
     let dummyAddress = Address ""
     let dummyOutput  = TxOut dummyAddress (TokenBundle.fromCoin deposit)
     let outs = dummyOutput :| []
-    selectAssets @ctx @s @k ctx w tx outs $ \s sel -> transform s $ sel
+    selectAssets @ctx @s @k ctx wal tx outs $ \s sel -> transform s $ sel
         { outputsCovered = mempty
         , changeGenerated =
             let
@@ -1318,15 +1319,13 @@ selectAssets
         , HasNetworkLayer ctx
         )
     => ctx
-    -> (WalletId, UTxOIndex, Wallet s, Set Tx)
+    -> (UTxOIndex, Wallet s, Set Tx)
     -> TransactionCtx
     -> NonEmpty TxOut
     -> (s -> SelectionResult TokenBundle -> result)
     -> ExceptT ErrSelectAssets IO result
-selectAssets ctx (_wid, utxo, cp, pending) tx outs transform = do
-    let s = getState cp
-
-    guardWithdrawal pending
+selectAssets ctx (utxo, cp, pending) tx outs transform = do
+    guardPendingWithdrawal
 
     pp <- liftIO $ currentProtocolParameters nl
     liftIO $ traceWith tr $ MsgSelectionStart utxo outs
@@ -1335,7 +1334,8 @@ selectAssets ctx (_wid, utxo, cp, pending) tx outs transform = do
         (calcMinimumCost tl pp tx)
         (initSelectionCriteria tl pp tx utxo outs)
     liftIO $ traceWith tr $ MsgSelectionDone sel
-    withExceptT ErrSelectAssetsSelectionError $ except (transform s <$> sel)
+    withExceptT ErrSelectAssetsSelectionError $ except $
+        transform (getState cp) <$> sel
   where
     nl = ctx ^. networkLayer
     tl = ctx ^. transactionLayer @k
@@ -1346,9 +1346,9 @@ selectAssets ctx (_wid, utxo, cp, pending) tx outs transform = do
     -- withdrawal is executed, the reward pot is empty. So, to prevent two
     -- transactions with withdrawals to go through (which will inevitably cause
     -- one of them to never be inserted), we warn users early on about it.
-    guardWithdrawal :: Set Tx -> ExceptT ErrSelectAssets IO ()
-    guardWithdrawal pending' = do
-        case Set.lookupMin $ Set.filter hasWithdrawal pending' of
+    guardPendingWithdrawal :: ExceptT ErrSelectAssets IO ()
+    guardPendingWithdrawal =
+        case Set.lookupMin $ Set.filter hasWithdrawal pending of
             Just pendingWithdrawal | withdrawalToCoin (txWithdrawal tx) /= Coin 0 ->
                 throwE $ ErrSelectAssetsAlreadyWithdrawing pendingWithdrawal
             _otherwise ->