--- a/lib/core/src/Cardano/Wallet/Api/Server.hs
+++ b/lib/core/src/Cardano/Wallet/Api/Server.hs
@@ -1196,7 +1196,7 @@ selectCoins ctx genChange (ApiT wid) body = do
         let transform = \s sel ->
                 W.assignChangeAddresses genChange sel s
                 & uncurry W.selectionToUnsignedTx
-        w <- liftHandler $ W.readUTxOIndex @_ @s @k wrk wid
+        w <- liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
         utx <- liftHandler
             $ W.selectAssets  @_ @s @k wrk w txCtx outs transform
 
@@ -1240,9 +1240,9 @@ selectCoinsForJoin ctx knownPools getPoolStatus pid wid = do
         let transform = \s sel ->
                 W.assignChangeAddresses (delegationAddress @n) sel s
                 & uncurry W.selectionToUnsignedTx
-        w <- liftHandler $ W.readUTxOIndex @_ @s @k wrk wid
+        wal <- liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
         utx <- liftHandler
-            $ W.selectAssetsNoOutputs @_ @s @k wrk w txCtx transform
+            $ W.selectAssetsNoOutputs @_ @s @k wrk wid wal txCtx transform
         (_, path) <- liftHandler
             $ W.readRewardAccount @_ @s @k @n wrk wid
 
@@ -1265,8 +1265,7 @@ selectCoinsForQuit
     -> Handler (Api.ApiCoinSelection n)
 selectCoinsForQuit ctx (ApiT wid) = do
     withWorkerCtx ctx wid liftE liftE $ \wrk -> do
-        action <- liftHandler
-            $ W.quitStakePool @_ @s @k @n wrk wid
+        action <- liftHandler $ W.quitStakePool @_ @s @k @n wrk wid
 
         (wdrl, _mkRwdAcct) <- mkRewardAccountBuilder @_ @s @k @n ctx wid Nothing
         let txCtx = defaultTransactionCtx
@@ -1277,11 +1276,10 @@ selectCoinsForQuit ctx (ApiT wid) = do
         let transform = \s sel ->
                 W.assignChangeAddresses (delegationAddress @n) sel s
                 & uncurry W.selectionToUnsignedTx
-        w <- liftHandler $ W.readUTxOIndex @_ @s @k wrk wid
+        wal <- liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
         utx <- liftHandler
-            $ W.selectAssetsNoOutputs @_ @s @k wrk w txCtx transform
-        (_, path) <- liftHandler
-            $ W.readRewardAccount @_ @s @k @n wrk wid
+            $ W.selectAssetsNoOutputs @_ @s @k wrk wid wal txCtx transform
+        (_, path) <- liftHandler $ W.readRewardAccount @_ @s @k @n wrk wid
 
         pure $ mkApiCoinSelection [] (Just (action, path)) utx
 
@@ -1438,7 +1436,7 @@ postTransaction ctx genChange (ApiT wid) body = do
             }
 
     (sel, tx, txMeta, txTime) <- withWorkerCtx ctx wid liftE liftE $ \wrk -> do
-        w <- liftHandler $ W.readUTxOIndex @_ @s @k wrk wid
+        w <- liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
         sel <- liftHandler
             $ W.selectAssets @_ @s @k wrk w txCtx outs (const Prelude.id)
         (tx, txMeta, txTime, sealedTx) <- liftHandler
@@ -1544,7 +1542,7 @@ postTransactionFee ctx (ApiT wid) body = do
             , txMetadata = getApiT <$> body ^. #metadata
             }
     withWorkerCtx ctx wid liftE liftE $ \wrk -> do
-        w <- liftHandler $ W.readUTxOIndex @_ @s @k wrk wid
+        w <- liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
         let runSelection = W.selectAssets @_ @s @k wrk w txCtx outs getFee
               where outs = coerceCoin <$> body ^. #payments
                     getFee = const (selectionDelta TokenBundle.getCoin)
@@ -1594,9 +1592,10 @@ joinStakePool ctx knownPools getPoolStatus apiPoolId (ApiT wid) body = do
                 , txTimeToLive = ttl
                 , txDelegationAction = Just action
                 }
-        w <- liftHandler $ W.readUTxOIndex @_ @s @k wrk wid
+        wal <- liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
         sel <- liftHandler
-            $ W.selectAssetsNoOutputs @_ @s @k wrk w txCtx (const Prelude.id)
+            $ W.selectAssetsNoOutputs @_ @s @k wrk wid wal txCtx
+            $ const Prelude.id
         (tx, txMeta, txTime, sealedTx) <- liftHandler
             $ W.signTransaction @_ @s @k wrk wid genChange mkRwdAcct pwd txCtx sel
         liftHandler
@@ -1631,15 +1630,15 @@ delegationFee
 delegationFee ctx (ApiT wid) = do
     withWorkerCtx ctx wid liftE liftE $ \wrk -> liftHandler $ do
         w <- withExceptT ErrSelectAssetsNoSuchWallet $
-            W.readUTxOIndex @_ @s @k wrk wid
+            W.readWalletUTxOIndex @_ @s @k wrk wid
         deposit <- W.calcMinimumDeposit @_ @s @k wrk wid
         mkApiFee (Just deposit) <$> W.estimateFee (runSelection wrk deposit w)
   where
     txCtx :: TransactionCtx
     txCtx = defaultTransactionCtx
 
-    runSelection wrk deposit w =
-        W.selectAssetsNoOutputs @_ @s @k wrk w txCtx calcFee
+    runSelection wrk deposit wal =
+        W.selectAssetsNoOutputs @_ @s @k wrk wid wal txCtx calcFee
       where
         calcFee _ = Coin.distance deposit . selectionDelta TokenBundle.getCoin
 
@@ -1676,9 +1675,10 @@ quitStakePool ctx (ApiT wid) body = do
                 , txDelegationAction = Just action
                 }
 
-        w <- liftHandler $ W.readUTxOIndex @_ @s @k wrk wid
+        wal <- liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
         sel <- liftHandler
-            $ W.selectAssetsNoOutputs @_ @s @k wrk w txCtx (const Prelude.id)
+            $ W.selectAssetsNoOutputs @_ @s @k wrk wid wal txCtx
+            $ const Prelude.id
         (tx, txMeta, txTime, sealedTx) <- liftHandler
             $ W.signTransaction @_ @s @k wrk wid genChange mkRwdAcct pwd txCtx sel
         liftHandler